pyproject.toml
[project]
name = "kmu-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.1.0",
    "dearpygui-ext>=2.0.0",
    "icecream>=2.1.8",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "pynput>=1.8.1",
    "python-dotenv>=1.2.1",
    "returns>=0.26.0",
    "sounddevice>=0.5.3",
]
